move ethereum init from cli to ethereum
