+++ src/chains/ethereum/ethereum/package.json
@@ -54,6 +54,7 @@
     "@ethereumjs/common": "2.6.5",
     "@ethereumjs/tx": "3.5.2",
     "@ethereumjs/vm": "5.9.3",
+    "@ganache/colors": "0.3.0",
     "@ganache/console.log": "0.2.0",
     "@ganache/ethereum-address": "0.5.2",
     "@ganache/ethereum-block": "0.5.3",
